refactored remaining /configs/get to use newer ConfigContext operations and removed now unused /configs/get endpoint
