+++ ui/desktop/src/components/settings_v2/mode/ModeSection.tsx
@@ -1,7 +1,7 @@
-import { useEffect, useState } from 'react';
-import { getApiUrl, getSecretKey } from '../../../config';
+import { useEffect, useState, useCallback } from 'react';
 import { all_goose_modes, ModeSelectionItem } from './ModeSelectionItem';
 import { View, ViewOptions } from '../../../App';
+import { useConfig } from '../../ConfigContext';
 
 interface ModeSectionProps {
   setView: (view: View, viewOptions?: ViewOptions) => void;
@@ -9,53 +9,32 @@ interface ModeSectionProps {
 
 export const ModeSection = ({ setView }: ModeSectionProps) => {
   const [currentMode, setCurrentMode] = useState('auto');
+  const { read, upsert } = useConfig();
 
   const handleModeChange = async (newMode: string) => {
-    const storeResponse = await fetch(getApiUrl('/configs/store'), {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'X-Secret-Key': getSecretKey(),
-      },
-      body: JSON.stringify({
-        key: 'GOOSE_MODE',
-        value: newMode,
-        isSecret: false,
-      }),
-    });
-
-    if (!storeResponse.ok) {
-      const errorText = await storeResponse.text();
-      console.error('Store response error:', errorText);
+    try {
+      await upsert('GOOSE_MODE', newMode, false);
+      setCurrentMode(newMode);
+    } catch (error) {
+      console.error('Error updating goose mode:', error);
       throw new Error(`Failed to store new goose mode: ${newMode}`);
     }
-    setCurrentMode(newMode);
   };
 
-  useEffect(() => {
-    const fetchCurrentMode = async () => {
-      try {
-        const response = await fetch(getApiUrl('/configs/get?key=GOOSE_MODE'), {
-          method: 'GET',
-          headers: {
-            'Content-Type': 'application/json',
-            'X-Secret-Key': getSecretKey(),
-          },
-        });
-
-        if (response.ok) {
-          const { value } = await response.json();
-          if (value) {
-            setCurrentMode(value);
-          }
-        }
-      } catch (error) {
-        console.error('Error fetching current mode:', error);
+  const fetchCurrentMode = useCallback(async () => {
+    try {
+      const mode = (await read('GOOSE_MODE', false)) as string;
+      if (mode) {
+        setCurrentMode(mode);
       }
-    };
+    } catch (error) {
+      console.error('Error fetching current mode:', error);
+    }
+  }, [read]);
 
+  useEffect(() => {
     fetchCurrentMode();
-  }, []);
+  }, [fetchCurrentMode]);
 
   return (
     <section id="mode" className="px-8">
